feature(service, dao) change int id to Long id

+++ src/main/java/by/netcracker/artemyev/dao/GeneralDao.java
@@ -10,5 +10,5 @@ public interface GeneralDao<T> {
     void add(T object) throws DaoException;
     void update(T object) throws DaoException;
     void remove(T object) throws DaoException;
-    T getById(int id) throws DaoException;
+    T getById(Long id) throws DaoException;
 }
