+++ src/main/java/by/netcracker/artemyev/service/EmployeeService.java
@@ -12,6 +12,6 @@ public interface EmployeeService extends GeneralService<Employee> {
     void add(Employee object) throws ServiceException;
     void update(Employee object) throws ServiceException;
     void remove(Employee object) throws ServiceException;
-    Employee getById(int id) throws ServiceException;
+    Employee getById(Long id) throws ServiceException;
     List<Employee> getAll() throws ServiceException;
 }
